pyproject.toml
[project]
name = "resume-cli"
version = "0.1.0"
description = "Agentic enabled tool that creates pretty resumes."
authors = [
    {name = "David C Reay", email = "david.c.reay@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "typer>=0.21.1,<0.22.0",
    "jinja2>=3.1.6,<4.0.0",
    "pyyaml>=6.0.3,<7.0.0",
]

[project.optional-dependencies]
tailor = [
    "openai>=1.0.0",
    "boto3>=1.34.0",
    "jsonschema>=4.0.0",
]
job-files = [
    "pypdf>=4.0.0",
    "python-docx>=1.0.0",
]
all = [
    "openai>=1.0.0",
    "boto3>=1.34.0",
    "jsonschema>=4.0.0",
    "pypdf>=4.0.0",
    "python-docx>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
resume-cli = "resume_cli.main:main"

[dependency-groups]
dev = [
    "isort>=7.0.0,<8.0.0",
    "pytest>=9.0.2,<10.0.0",
    "pytest-cov>=7.0.0,<8.0.0",
    "pytest-env>=1.3.2,<2.0.0",
    "pylint>=4.0.4,<5.0.0",
    "flake8>=7.3.0,<8.0.0",
    "ruff>=0.15.1,<0.16.0",
    "pytest-mock>=3.15.1,<4.0.0",
    "pre-commit>=4.0.0",
]

[tool.uv]
default-groups = ["dev"]

[tool.pytest.ini_options]
markers = [
    "unit: mark a test as a unit test",
    "login: tests login features",
    "configure: tests configure features",
    "billing",
    "reporting"
]

[tool.ruff]
line-length = 150

[tool.pylint."FORMAT"]
max-line-length = 150

[tool.flake8]
max-line-length = 150
